Modifier rethink, and knock-on effects
With this change, all modifiers except `GUI` are on the left hand, and
all layer shifts on the right hand.  This means that all left-handed
keys are game-friendly, removing my need for a separate gaming layer.

To compensate for the loss of extra keys in the ‘gaming input’ layer,
I have added a ‘mirrored’ layer to grant access to the other half of
the keyboard when operating one-handed.

`Shift` being the centre key on the left hand makes it easy to combine
with both `R` and also `Ctrl` and `Alt`, which are also easy to
combine with each other since they're triggered with the same finger.

The control layer and the `GUI` modifier are likewise paired on the
right hand, making them easy to combine as I typically do when
controlling my window manager.

I could consider moving Fn and media keys to the unused half of the
mirror layout.

I've also rearranged some symbols for command-line and Rust usage:

- Rust uses `_` much more than `-`, so I have used overrides to swap
  the shifted and unshifted states of that key

- I use the digraphs `..` and `./` a lot on the command line and find
  north-north combinations awkward to type, so those keys have been
  moved to inward laterals